Add files via upload
added features: completed the histogram, matchHistograms and equalizeHistogram functions

histogram - shows the frequency of each pixel intensity, gives an insight into the contrast and brightness of the image

equalizeHistogram- improves the contrat of an image by redistributing pixel intensity values, making the histogram more spread out. Results in better visibility of features having low contrast

matchHistograms - adjusts the pixel values of source image according to reference image to match the histogram of the reference image. Which makes the source image appear more similar to the reference in terms of brightness and contrast